+++ src/SHMultiTaperSE.f95
@@ -75,14 +75,14 @@ subroutine SHMultiTaperSE(mtse, sd, cilm, lmax, tapers, taper_order, lmaxt, &
 
     real(dp), intent(out) :: mtse(:), sd(:)
     real(dp), intent(in) :: cilm(:,:,:), tapers(:,:)
-    integer, intent(in) :: lmax, lmaxt, K, taper_order(:)
+    integer(int32), intent(in) :: lmax, lmaxt, K, taper_order(:)
     real(dp), intent(in), optional :: alpha(:), lat, lon, taper_wt(:)
-    integer, intent(in), optional :: csphase, norm
-    integer, intent(out), optional :: exitstatus
-    integer :: i, l, phase, mnorm, astat(7), lmaxmul, nlat, nlong
+    integer(int32), intent(in), optional :: csphase, norm
+    integer(int32), intent(out), optional :: exitstatus
+    integer(int32) :: i, l, phase, mnorm, astat(7), lmaxmul, nlat, nlong
     real(dp) :: se(lmax-lmaxt+1, K), x(3), pi, factor
     real(dp), allocatable, save :: zero(:), w(:)
-    integer, save :: first = 1, lmaxmul_last = -1
+    integer(int32), save :: first = 1, lmaxmul_last = -1
     real(dp), allocatable :: shwin(:,:,:), shloc(:,:,:), dj(:,:,:), &
                              shwinrot(:,:,:), grid1glq(:,:), gridwinglq(:,:), &
                              temp(:,:)
